Complete publish support for AWS
The publishing procedure on AWS includes four phases: upload_to_container,
import_snapshot, register_image and share_image. These methods take a
publishing metadata object for simplicity. The benefit is the caller does not
have to understand this procedure. All they need to do is supply a metadata
object to publish and everything should be taken care of.