snapshot object has no name attr [RHELDST-11233]
QE Testing found an issue with snapshot sharing. When trying to log the
fact the snapshot is being shared, I tried using a "name" attribute like
when sharing an image, but the snapshot object doesnt have it. This
threw an exception, which was caught in pubtools_ami and the publush
function is then retried. On retry, the snapshot exists already, so the
snapshot_share function doesn't get triggered.