Allow AWSService to use tags
This commit changes the AWSService to allow it using tags to search for
an EC2 image as well as to tag new images after registering them.

IT's important to note that this change let the tags to be optional, it
is, it won't break existing workflows without tags.

Tagging is a requirement for the new marketplace upload workflows thus
it's being added to AWSService in a retrocompatible way.