Add 'AMI image delete' [RHELDST-7515]
This change introduces public methods for AWS service that
are capable of removing AWS image and snapshot.

The caller is expected to use AWSService.delete()
with proper metadata provided. Id of the image we want
to delete is the primary identifier, other metadata are
used as helpers is some edge cases.

AMI image is deregistered and its snapshot is deleted as long as they
exist in AWS.