Upload decompressed LZMA file for Azure
This commit patches the `upload_to_container` method from `AzureService`
to automatically detect LZMA compressed VHD files and send the
uncompressed data for upload when mounting the storage.

This is required for Azure as it can't handle VM images from compressed
files in its marketplace as well as it doesn't provide any API to
decompress a file already uploaded to its storage, leaving the only
alternative to upload the uncompressed blob.