Drop support of uploading by URL on Azure
The `blob_client.upload_blob_from_url` expects to receive a public URL
in order to request Azure to retrieve the content from it. This is an
undesired behavior as we're not using public facing URLs.

This commit drops the support of uploading by URL for Azure as it's an
unused feature which differs from the original goal.