Add ability to specify bootmode [RHELDST-18235]
From RHEL 8.9 and RHEL 9.3 onwards UEFI will be available. The boot mode
must be specified for AMI images to make use of the change. This change
adds the boot mode to the metadata and image sharing call. I've kept the
value as a boolean until this point since AWS requires specific string
values for different boot types.