Allow adding tags for snapshots and S3 objects
This commit allows `cloudimg` to add tags on EC2 snapshots as well as S3
objects.

Tags on images could be used for various use-cases in the future
e.g. will be useful to precisely determine to which product a particular
S3 object, snapshot or AMI belongs to, which will allow product-based
access control on those.

Refers to EXDSP-2299