Allow disabling the search for tags fallback
When searching for the image existence, by defaut, the library first
search it from its name and if not found by the tags. However, there are
certain situations which the latter search (tags) may not be executed.

This commit preserves the default experience of the library while
allowing to disable the tags search fallback wheenver the new publishing
metadata attribute `search_tags` is set to `False`.

Refers to SPSTRAT-451